% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AGCA4extremes-package}
\alias{AGCA4extremes}
\alias{AGCA4extremes-package}
\title{AGCA4extremes}
\description{
Anchored geodesic component analysis for multivariate extremes.
}
\author{
\strong{Maintainer}: Alberto Quaini \email{alberto91quaini@gmail.com}

Authors:
\itemize{
  \item Alberto Quaini \email{alberto91quaini@gmail.com}
}

}
\keyword{internal}
