% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agca.R
\name{agca_reconstruct}
\alias{agca_reconstruct}
\title{Reconstruct angular directions from leading AGCA components}
\usage{
agca_reconstruct(fit, p = fit$p)
}
\arguments{
\item{fit}{An object returned by \code{\link[=agca]{agca()}} or \code{\link[=agca_fit_directions]{agca_fit_directions()}}.}

\item{p}{Reconstruction rank.}
}
\value{
A matrix of reconstructed angular directions.
}
\description{
Reconstruct angular directions from leading AGCA components
}
