% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agca.R
\name{agca_variation_explained}
\alias{agca_variation_explained}
\title{Anchored variation explained}
\usage{
agca_variation_explained(fit)
}
\arguments{
\item{fit}{An object returned by \code{\link[=agca]{agca()}} or \code{\link[=agca_fit_directions]{agca_fit_directions()}}.}
}
\value{
Cumulative anchored variation explained by each rank.
}
\description{
Anchored variation explained
}
