% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{canonical_anchor}
\alias{canonical_anchor}
\title{Canonical anchor}
\usage{
canonical_anchor(d)
}
\arguments{
\item{d}{Ambient dimension.}
}
\value{
The balanced direction \eqn{d^{-1/2}(1,\ldots,1)}.
}
\description{
Canonical anchor
}
