% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.agca_bootstrap}
\alias{plot.agca_bootstrap}
\title{Plot AGCA bootstrap summaries}
\usage{
\method{plot}{agca_bootstrap}(x, statistic = c("variation_explained", "residual_risk"), ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=bootstrap_agca]{bootstrap_agca()}}.}

\item{statistic}{Statistic to plot.}

\item{...}{Additional graphical arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot AGCA bootstrap summaries
}
