% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{row_norms}
\alias{row_norms}
\title{Row Euclidean norms}
\usage{
row_norms(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A numeric vector containing one Euclidean norm per row.
}
\description{
Row Euclidean norms
}
