% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summary.agca_bootstrap}
\alias{summary.agca_bootstrap}
\title{Summarize AGCA bootstrap output}
\usage{
\method{summary}{agca_bootstrap}(object, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{An object returned by \code{\link[=bootstrap_agca]{bootstrap_agca()}}.}

\item{probs}{Quantile probabilities.}

\item{...}{Unused.}
}
\value{
A data frame of bootstrap summaries by rank.
}
\description{
Summarize AGCA bootstrap output
}
